rename operation file
